Moved my objects to objects.py
To avoid having to import one thousand things at the beginning of every script. When you want to use these objects, just use
from stcad.source_dev.objects import *
